+++ src/demos/Demos.Collisions.Interactable/Services/States/CollisionScenarioState.cs
@@ -1,4 +1,4 @@
-﻿using CollisionFormats.Serialization;
+﻿using CollisionFormats.Model;
 using Detach.Collisions;
 using System.Reflection;
