+++ src/internals/CollisionFormats/Model/CollisionAlgorithmParameter.cs
@@ -1,3 +1,3 @@
-﻿namespace CollisionFormats.Serialization;
+﻿namespace CollisionFormats.Model;
 
 public sealed record CollisionAlgorithmParameter(string TypeName, string Name);
